fix: refactor and fix hooks
